Increasing locality for constants
